backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# Database — PostgreSQL
sqlalchemy==2.0.30
psycopg2-binary==2.9.9   # PostgreSQL driver
alembic==1.13.1           # DB migrations (schema versioning)

# Data processing
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4

# AI / Audio
anthropic>=0.95.0          # Claude API (audio insight extraction)
openai-whisper==20231117   # Local audio transcription (requires ffmpeg)

# Utilities
pydantic==2.7.1
python-dotenv==1.0.1
